requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==3.0.4
Django==3.2
django-environ==0.4.5
gunicorn==20.1.0
idna==2.10
mysqlclient==2.0.3
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.25.11
